isaspec: generate `FpuRRR` addition spec (#201)
Generate spec for the `FpuRRR` add opcode.

This introduces a new approach for handling primitive functions in ASLp,
which have previously been mapped to spec expressions entirely in Rust
code. The `FPAdd` primitive is complex, so it is better to allow the
user to implement it in ISLE code itself. Therefore the primitive call
in ASLp is translated to a spec macro call in ISLE, and the macro
implementation is in a new `primitives.isle` file (currently `FPAdd` is
a placeholder to be filled out later). This approach may make sense for
other primitives such as the memory reads/writes.

Updates #198