verifier: invert meaning of priority tag (#204)
This PR modifies the meaning of the priority verification attribute.
Previously the attribute meant "negate all the higher priority
overlapping rules above me". Now it means, "negate me in all the
overlapping lower priority rules below me". Therefore it is a more
fine-grained configuration mechanism, and by default more conservative.

This is an initial response to the priority shadowing problem discussed
in #128. Specifically, if a higher priority rule defines an imprecise
specification of its match conditions that is an over-approximation,
then when negated it causes under-approximation of when the lower
priority rule matches. That can cause the verifier to incorrectly deduce
a lower priority rule is inapplicable. Inapplicability may be noticed by
an observant user, but a more pernicious version of the problem might
not cause inapplicability but still missed bugs.

With this change the default behavior is ignoring higher priority rules
(therefore over-approximation), and only bringing in negations
one-by-one.

Updates #128