aarch64: verify `udiv` (#221)
Verify `udiv` rules.

This approach improves on the prior verificaiton of divide instructions,
since we can now use state to model traps. We do so by introducing two
flags: one for an expected trap in CLIF, and a second for a trap on the
execution side. The post-condition for `lower` is updated to require
that a trap must happen on both sides or neither.

Updates #34 #49