aarch64: specs for verification of broken `sdiv` lowering (#233)
Attempt verification of the `sdiv` rule, which is relatively complex
given:

* Two conditions under which it can trap: divide by zero, and the
unrepresentable signed case `(-2^{N-1})/(-1)`.
* A delicate sequence of instructions to perform the checks, depending
on flags behavior.
* Multiple instructions in an expansion modifying the same state,
therefore the first use of conditional state modification #232.

This PR adds specs required to verify this rule, and has uncovered an
apparent bug! The overflow check used to detect the unrepresentable case
only works in the 32/64-bit cases, and breaks in 8/16. The bug is filed
as https://github.com/bytecodealliance/wasmtime/issues/9537.

In order to land the specs and proceed the rule has been marked with the
`broken` tag, and `broken` cases excluded from the CI verification run.

Updates #34 #49