type inference: only record choices with multiple branches (#266)
Verifier logs are full of type instantiations, and many of them are not
informative. There are cases where we had to create instantations for
type inference to work, but there's only one choice. For example, for
the `AMode` variants. This PR changes the forking logic so choices are
only recorded if they were for a non-trivial branch.